requirements.txt
Django==1.11.16 # pyup: <2.0
httplib2==0.12.0
Markdown==2.6.11 # pypu: <3.0.0
smartypants==2.0.1
psycopg2==2.7.6.1
olefile==0.46
Pillow==5.3.0
anyjson==0.3.3
billiard==3.5.0.5

vine==1.1.4
amqp==2.3.2

amqplib==1.0.2
kombu==3.0.37 # pyup: <4.0.0
django-picklefield==1.1.0
pyparsing==2.3.0
six==1.11.0
pytz==2018.7
python-dateutil==2.7.5
librabbitmq==2.0.0

celery==3.1.26.post2 # pyup: <4.0.0

oauth2client==4.0.0 # pyup: <4.1
pyasn1==0.4.4
pyasn1_modules==0.2.2
rsa==4.0
uritemplate==3.0.0
google-api-python-client==1.7.5
enum34==1.1.6; python_version < '3.4'
asn1crypto==0.24.0
cryptography==2.4.2
ecdsa==0.13
cffi==1.11.5
idna==2.7
bcrypt==3.1.4
pynacl==1.3.0
paramiko==2.4.2
raven==6.9.0
pyOpenSSL==18.0.0
ndg-httpsclient==0.5.1

chardet==3.0.4
idna==2.7
urllib3==1.24.1

requests==2.20.1
statsd==3.3.0
coverage==4.5.2
pep8==1.7.1
pyflakes==2.0.0
mccabe==0.6.1
pycodestyle==2.4.0
configparser==3.5.0
flake8==3.6.0
logilab-common==1.4.2
logilab-astng==0.24.3
editdistance==0.5.2
pylint==1.9.3 # pyup: <2.0.0
ldap3==2.5.1
surelink==1.0.1
boto==2.49.0
traitlets==4.3.2
ipython_genutils==0.2.0
decorator==4.3.0
ptyprocess==0.6.0
pexpect==4.6.0
pickleshare==0.7.5
simplegeneric==0.8.1
path.py==11.5.0

ipython==5.8.0 # pyup: <6.0.0

sqlparse==0.2.4
filechunkio==1.8
simplejson==3.16.0
httpretty==0.9.4 # pyup: <0.9.5
contextlib2==0.5.5
ipaddress==1.0.22
text-unidecode==1.2  # for faker
Faker==1.0.0
factory_boy==2.11.1
rcssmin==1.0.6
rjsmin==1.0.12

djangowind==1.2.1
django-staticmedia==0.2.2
django-celery==3.2.2
requirements/src/django_oembed-0.1.4ctl-py2.py3-none-any.whl
django-taggit==0.23.0
django-statsd-mozilla==0.4.0
django-smoketest==1.1.0
django-waffle==0.15.0
django-debug-toolbar==1.10.1
django-jenkins==0.110.0
django-appconf==1.0.2
django-compressor==2.2
django-extensions==2.1.4
typing==3.6.6 # needed for django-extensions
django-markwhat==1.6.1
tornado==4.5.3 # pyup: <5.0
django-storages==1.7.1
django-cacheds3storage==0.1.2
gunicorn==19.9.0
django-impersonate==1.3
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-s3sign==0.1.4
django-smtp-ssl==1.0
certifi==2018.11.29

ccnmtlsettings==1.3.0 # pyup: <1.4.0

pbr==5.1.1
PyYAML>=3.10.0 # MIT
stevedore>=1.20.0 # Apache-2.0
bandit==1.5.1

defusedxml==0.5.0
lxml==4.2.5
requests-toolbelt==0.8.0
isodate==0.6.0
cached-property==1.5.1
zeep==2.5.0 # pyup: <3.0.0
pypanopto==0.0.5
django-bootstrap4==0.0.7

